Assign AnonID in Excel instead.
Prefix and Digits now dropped from the sheet. These are instead
taken from the AnonID on first Row, eg MYPROJ-0000 gives Prefix
MYPROJ- and Digits 4.

Assigning AnonIDs in Excel is completely fine if Persons are
required unique, because this allows user to type out the first
AnonID and then drag down to assign the rest sequentially.